In Assignment 4 we are supposed to:
Create a side-by-side boxplot (boxplot(x, ...)) and histogram ((hist(x, ...)).

Discuss the outcome of your results regarding patients BPs & MD’s Ratings.

In order to create a side-by-side boxplot of the frequence and blood pressure I could have used the very simple function call:
boxplot(Freq, bloodp).
Since the values of frequency and blood pressure are on a different scale (Frequency between 0 and 1, blood pressure between 32 and 205) this side by side boxplot would not have been very visually clear and therefore would not have provided much information.
I therefore opted to produce two separate boxplots with their own scale that are displayed on the screen side by side with their corresponding histograms using the  par(mfrow=c(2,2)) function. This function call creates a "matrix" of graphs with 4 separates graphs, arranged in 2 rows and 2 columns.
The histogram and box plot of the blood pressure show, that the blood pressure measured falls into the middle section of all occurrences more often than into the extremes.
In order to then also have a visual display of the first, second and final decision while not "overwriting" the previous graphs I decided to open a new window using  the windows() function. In this window I created a matrix for 3 graphs, positioned underneath of each other, displaying bar plots of the first, second and final decision in order to compare the outcomes.
With these three graphs aligned underneath of each other it is easily visible that only in 2 out of the 10 occurrences the first, second and final decision are in agreement.
With all graphs displayed in 2 separate windows containing several graphs each, the user can look at graphs of all given data at the same time.